refactor: reorganize FailedPreconditionExceptions
